+++ stl/CHANGELOG.md
@@ -193,5 +193,6 @@
 - Added `operator >=` to `Array` class.
 
 ## 0.3.2
-- Fix upload issue to- RELATED to project
+- Fix upload issue to-RELATED to project
 - applied the @override annotation to the operator + method in array.dart
+- Other minor fixes related to dart analyzer test